requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
requests==2.31.0
PyYAML==6.0.1
Jinja2==3.1.2
python-multipart==0.0.6
authlib==1.3.0
cryptography==41.0.7
kafka-python==2.0.2
